Split destructive DDL doctor tests into their own async: false module
test/mix/tasks/phoenix_kit_doctor_orphaned_fk_test.exs (async: true) ran
three tests that execute real DDL against phoenix_kit_activities and
phoenix_kit_user_oauth_providers, including ALTER TABLE ... ADD
CONSTRAINT ... REFERENCES phoenix_kit_users. That statement takes a
ShareRowExclusiveLock on both the altered and the referenced table --
and phoenix_kit_users is written to by well over a dozen other async
test files. A live full-suite run deadlocked (Postgres 40P01) between
one of these ALTERs and an unrelated async test's write to
phoenix_kit_users. Passing in isolation proved nothing: the race only
exists when the whole suite's async tests are actually running
concurrently, which a single-file run never exercises.

Moved all three DDL tests into a new module,
DoctorOrphanedFkDestructiveTest, async: false. ExUnit runs every
async: false module only after every async: true module has finished,
so by the time any test here runs nothing else in the suite is still
touching phoenix_kit_users -- the deadlock is structurally unreachable,
not just less likely. The read-only tests that shared a describe block
with a moved destructive test stay behind in the original async file.
